requirements.txt
langchain
langchain-openai
langchain-core
langchain-community
langchain-experimental
langchain-huggingface
langchain-text-splitters
requests
unstructured
chromadb
bs4
protobuf==3.20.*
flask
sentence_transformers
torch==2.4.0
google-generativeai
python-dotenv
langchain_google_genai
faiss-cpu